+++ changelog.md
@@ -6,6 +6,7 @@ Minor fixes.
 
 AssetView:
 - Support for renaming items.
+- Support for `options.onItemDragged`.
 - Support select and preview also from the asset explorer.
 - `AssetView.addItem` (previously `_addItem`) now updates the tree using optional `updateTree` param.
 - Fixed some empty folder path issues.